Fix segfault when run on python 3.7
Problem arrose due to difference in pointer size to float arrays
with 3.7 appearing to use 64 bit pointers while 3.5 was using 32
bit pointers.
In the original call to libtheofast.heat_potential_time_step() the
pointers were just being passed as eg h_pot.ctypes.data which were
being interpreted as 32 bit ints in the function signature leading
to the pointers being trucated and then upcast again in the C code.
This resulted in pointers in the C domain looking like:
0xffffffff85b70aa0
and producing seg faults as soon as the pointers were dereferenced.
Therefore, casting the input pointers correctly as:
ctypes.c_void_p(h_pot.ctypes.data)
avoided the downcasting and preserved the accuracy of the pointer
addresses.
No changes were required to the C code to fix the issue.